Correct the output of REST in order to use generics where possible. Modify the python clients.

+++ CondDBWeb/src/main/java/conddb/web/controllers/GlobalTagMapExpRestController.java
@@ -10,6 +10,7 @@ import javax.ws.rs.DELETE;
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
+import javax.ws.rs.Produces;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
@@ -31,6 +32,7 @@ import conddb.web.exceptions.ConddbWebException;
 import conddb.web.resources.GlobalTagMapResource;
 import conddb.web.resources.Link;
 import conddb.web.resources.SpringResourceFactory;
+import conddb.web.resources.generic.GenericPojoResource;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
@@ -55,6 +57,7 @@ public class GlobalTagMapExpRestController extends BaseController {
 	
 	@POST
 	@Consumes({ MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML })
+	@Produces({ MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML })
 	@Path(Link.GLOBALTAGMAPS)
 	@ApiOperation(value = "Create a GlobalTagMap.",
     notes = "Input data are in json, and should match all needed fields for a new global tag to tag association.\n"
@@ -64,13 +67,10 @@ public class GlobalTagMapExpRestController extends BaseController {
 
 		try {
 			log.info("Request for creating a new global tag mapping using data map of size "+map.size());
-
 			GlobalTagMap globaltagmap = createGlobalTagMap(map);
+			
 			GlobalTagMap saved = globalTagService.insertGlobalTagMap(globaltagmap);
-			String resid = new String(saved.getId().toString());
-			saved.setResId(resid);
-			GlobalTagMapResource resource = (GlobalTagMapResource) springResourceFactory.getResource("globaltagmap", info,
-					saved);
+			GenericPojoResource<GlobalTagMap> resource = (GenericPojoResource) springResourceFactory.getGenericResource(info, saved, 0, null);
 			return created(resource);
 		} catch (ConddbServiceException e) {
 			String msg = "Error creating association resource using "+map.toString();
@@ -81,6 +81,7 @@ public class GlobalTagMapExpRestController extends BaseController {
 	@Path(Link.GLOBALTAGMAPS+"/{id}")
 	@POST
 	@Consumes(MediaType.APPLICATION_JSON)
+	@Produces({ MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML })
 	@ApiOperation(value = "Update a GlobalTagMap association.",
     notes = "Input data are in json, and should match the fields that can be updated: record and label.",
     response=GlobalTagMap.class)
@@ -88,9 +89,8 @@ public class GlobalTagMapExpRestController extends BaseController {
 			@ApiParam(value = "id: id of the globaltagmap to be updated", required = true) 
 			@PathParam("id") Long id, Map map)
 			throws ConddbWebException {
-		Response resp;
 		try {
-			log.info("Request for updating global tag "+id+" using "+map.size());
+			log.info("Request for updating global tag mapping "+id+" using "+map.size());
 			GlobalTagMap existing = globalTagService.getGlobalTagMap(id);
 			if (existing == null) {
 				String msg = "Error updating association resource: id "+id+" not found!";
@@ -102,9 +102,8 @@ public class GlobalTagMapExpRestController extends BaseController {
 			if (map.containsKey("record")) {
 				existing.setRecord(String.valueOf(map.get("record")));
 			}
-			existing = globalTagService.insertGlobalTagMap(existing);
-			resp = Response.ok(new GlobalTagMapResource(info, existing), MediaType.APPLICATION_JSON).build();
-			return resp;
+			GenericPojoResource<GlobalTagMap> resource = (GenericPojoResource) springResourceFactory.getGenericResource(info, existing, 0, null);
+			return ok(resource);
 		} catch (ConddbServiceException e) {
 			log.debug("Generate exception using an ConddbService exception..."+e.getMessage());
 			String msg = "Error updating association resource: internal server exception !";
@@ -113,23 +112,30 @@ public class GlobalTagMapExpRestController extends BaseController {
 	}
 
 	@Path(Link.GLOBALTAGMAPS+"/{id}")
+	@Produces({ MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML })
 	@DELETE
-	@ApiOperation(value = "Delete a globaltag.",
+	@ApiOperation(value = "Delete a globaltag mapping.",
     notes = "It should be used one global tag at the time. This method is meant for administration purposes.",
     response=GlobalTagMap.class)
-	public Response deleteGlobalTagMap(
+	public Response deleteGlobalTagMap(@Context UriInfo info, 
 			@ApiParam(value = "id: id of the association to be deleted", required = true) 
 			@PathParam("id") Long id) throws ConddbWebException {
-		Response resp;
-		GlobalTagMap existing = null;
 		try {
-			existing = globalTagAdminService.deleteGlobalTagMap(id);
+			GlobalTagMap existing = globalTagService.getGlobalTagMapFetchChildren(id);
 			if (existing == null) {
 				String msg = "Error removing association resource: id "+id+" not found!";
 				throw buildException(msg, msg, Response.Status.NOT_FOUND);
 			}
-			resp = Response.ok(existing).build();
-			return resp;
+			GlobalTag gtag = existing.getGlobalTag();
+			if (gtag.islocked()) {
+				log.debug("Cannot remove a mapping for a locked global tag: "+gtag.getName());
+				String msg = "Error removing an association resource: global tag is locked !";
+				throw buildException(msg, msg, Response.Status.PRECONDITION_FAILED);				
+			}
+			existing = globalTagAdminService.deleteGlobalTagMap(id);
+			GenericPojoResource<GlobalTagMap> resource = (GenericPojoResource) springResourceFactory.getGenericResource(info, existing, 0, null);
+
+			return ok(resource);
 		} catch (ConddbServiceException e) {
 			log.debug("Generate exception using a ConddbService exception..."+e.getMessage());
 			String msg = "Error removing an association resource: internal server exception !";
@@ -145,10 +151,16 @@ public class GlobalTagMapExpRestController extends BaseController {
 		}
 		if (map.containsKey("globaltagname")) {
 			GlobalTag globaltag = globalTagService.getGlobalTag((String.valueOf(map.get("globaltagname"))));
+			if (globaltag == null) {
+				throw new ConddbServiceException("Global tag "+String.valueOf(map.get("globaltagname"))+" does not exists ");
+			}
 			globaltagmap.setGlobalTag(globaltag);
 		}
 		if (map.containsKey("tagname")) {
 			Tag tag = globalTagService.getTag((String.valueOf(map.get("tagname"))));
+			if (tag == null) {
+				throw new ConddbServiceException("Tag "+String.valueOf(map.get("tagname"))+" does not exists ");
+			}
 			globaltagmap.setSystemTag(tag);
 		}
 		if (map.containsKey("record")) {
